+++ apiphany-httpclient5/src/test/java/org/apiphany/client/http/ApacheHC5PropertiesTest.java
@@ -18,6 +18,8 @@ import com.fasterxml.jackson.dataformat.yaml.YAMLFactory;
 
 /**
  * Test class for {@link ApacheHC5Properties}.
+ *
+ * @author Radu Sebastian LAZIN
  */
 class ApacheHC5PropertiesTest {